updated python question option and refactor questions architecture
