--- app/src/main/java/com/better/alarm/model/QuestionType.kt
@@ -1,7 +1,0 @@
-package com.better.alarm.model
-
-enum class QuestionType {
-    ALL,
-    JAVA,
-    SQL
-}
