--- app/src/main/java/com/better/alarm/model/Questions.kt
@@ -1,3 +1,0 @@
-package com.better.alarm.model
-
-
