+++ app/src/main/java/com/better/alarm/question/QuestionType.kt
@@ -1,0 +1,8 @@
+package com.better.alarm.question
+
+enum class QuestionType {
+    ALL,
+    JAVA,
+    SQL,
+    PYTHON
+}
